test(fft): correct CUDA RED coverage
