+++ crates/tenferro-fft/src/cuda/tests/ffi.rs
@@ -1,10 +1,12 @@
+use std::error::Error;
 use std::ffi::{OsStr, OsString};
 
 use super::super::error::CudaFftError;
 use super::super::ffi::{
-    cufft_library_candidates, map_cufft_status, CUFFT_ALLOC_FAILED, CUFFT_EXEC_FAILED,
-    CUFFT_INTERNAL_ERROR, CUFFT_INVALID_PLAN, CUFFT_INVALID_SIZE, CUFFT_INVALID_TYPE,
-    CUFFT_INVALID_VALUE, CUFFT_SETUP_FAILED, CUFFT_SUCCESS, CUFFT_UNALIGNED_DATA,
+    cufft_library_candidates, map_cufft_status, CufftLibrary, CUFFT_ALLOC_FAILED,
+    CUFFT_EXEC_FAILED, CUFFT_INTERNAL_ERROR, CUFFT_INVALID_PLAN, CUFFT_INVALID_SIZE,
+    CUFFT_INVALID_TYPE, CUFFT_INVALID_VALUE, CUFFT_SETUP_FAILED, CUFFT_SUCCESS,
+    CUFFT_UNALIGNED_DATA,
 };
 
 #[test]
@@ -34,6 +36,20 @@ fn library_candidates_use_defaults_when_override_is_absent_or_empty() {
     ];
     assert_eq!(cufft_library_candidates(None), expected);
     assert_eq!(cufft_library_candidates(Some(OsStr::new(":"))), expected);
+}
+
+#[test]
+fn candidate_only_load_failure_preserves_typed_source_without_process_environment() {
+    let candidate = OsString::from("/definitely/missing/libcufft-for-unit-test.so");
+    let error = match CufftLibrary::load_from_paths_for_tests(vec![candidate.clone()]) {
+        Ok(_) => panic!("a missing candidate must not load"),
+        Err(error) => error,
+    };
+    assert!(matches!(
+        &error,
+        CudaFftError::LibraryLoad { paths, .. } if paths == "/definitely/missing/libcufft-for-unit-test.so"
+    ));
+    assert!(error.source().is_some(), "loader source must be retained");
 }
 
 #[test]
